Refactored positionContents into layoutContent_ and polished implementation of block adjustment in flyouts.
