+++ core/flyout_base.js
@@ -17,7 +17,6 @@ goog.require('Blockly.Block');
 goog.require('Blockly.blockRendering');
 goog.require('Blockly.browserEvents');
 goog.require('Blockly.ComponentManager');
-goog.require('Blockly.constants');
 goog.require('Blockly.DeleteArea');
 goog.require('Blockly.Events');
 /** @suppress {extraRequire} */
@@ -124,14 +123,14 @@ Blockly.Flyout = function(workspaceOptions) {
   /**
    * List of gaps between flyout content items (blocks, buttons, etc).
    * @type {!Array<number>}
-   * @private
+   * @protected
    */
   this.gaps_ = [];
 
   /**
    * List of the contents of the flyout (blocks, buttons, etc).
    * @type {!Array<!Blockly.FlyoutButton|!Blockly.BlockSvg>}
-   * @private
+   * @protected
    */
   this.contents_ = [];
 
@@ -395,95 +394,6 @@ Blockly.Flyout.prototype.getFlyoutScale = function() {
  */
 Blockly.Flyout.prototype.getWorkspace = function() {
   return this.workspace_;
-};
-
-/**
- * Respositions the blocks, buttons and labels in the flyout based on their
- * current dimensions and order.
- * @param {!Blockly.constants.AXIS} primaryAxis The axis to position content
- * along.
- */
-Blockly.Flyout.prototype.positionContents = function(primaryAxis) {
-  this.workspace_.setResizesEnabled(false);
-  this.reflow();
-  var margin = this.MARGIN;
-  var cursorX = margin + this.tabWidth_;
-  var cursorY = margin;
-
-  var contents = Array.from(this.contents_);
-  var gaps = Array.from(this.gaps_);
-  // For RTL in a X axis layout, if the length of the blocks is less than the
-  // width of the flyout, the blocks need to be offset so that they appear
-  // right-aligned. To do so, we calculate the width of the blocks (and offset,
-  // if any) before adjusting the blocks' positions.
-  if (this.RTL && primaryAxis === Blockly.constants.AXIS.X) {
-    contents.reverse();
-    gaps.reverse();
-
-    var contentWidth = cursorX;
-    for (var i = 0, item; (item = contents[i]); i++) {
-      if (item instanceof Blockly.BlockSvg) {
-        contentWidth += item.getHeightWidth().width;
-      } else if (item instanceof Blockly.FlyoutButton) {
-        contentWidth += item.width;
-      }
-      contentWidth += gaps[i];
-    }
-    var delta = this.getWidth() - contentWidth;
-    if (delta > 0) {
-      cursorX += delta;
-    }
-  }
-
-  // Delete all the event listeners.
-  for (var i = 0, listen; (listen = this.listeners_[i]); i++) {
-    Blockly.browserEvents.unbind(listen);
-  }
-  
-  for (var i = 0, item; (item = contents[i]); i++) {
-    if (item instanceof Blockly.BlockSvg) {
-      var block = item;
-      var root = block.getSvgRoot();
-      var dimensions = block.getHeightWidth();
-      var position = block.getRelativeToSurfaceXY();
-
-      var destinationX = cursorX;
-      if (block.outputConnection && !this.RTL) {
-        destinationX -= this.tabWidth_;
-      }
-
-      if (this.RTL && primaryAxis === Blockly.constants.AXIS.X) {
-        destinationX += dimensions.width;
-      } else if (this.RTL && primaryAxis === Blockly.constants.AXIS.Y) {
-        destinationX = (this.width_ / this.workspace_.scale) - (margin * 2) +
-            (block.outputConnection ? this.tabWidth_ : 0);
-      }
-      var destinationY = cursorY;
-
-      var deltaX = destinationX - position.x;
-      var deltaY = destinationY - position.y;
-
-      if (deltaX || deltaY) {
-        block.moveBy(deltaX, deltaY);
-      }
-      this.moveRectToBlock_(block.flyoutRect_, block);
-      this.addBlockListeners_(root, block, block.flyoutRect_);
-    } else if (item instanceof Blockly.FlyoutButton) {
-      item.moveTo(cursorX, cursorY);
-      var dimensions = item;
-    }
-
-
-    switch (primaryAxis) {
-      case Blockly.constants.AXIS.X:
-        cursorX += dimensions.width + gaps[i];
-        break;
-      case Blockly.constants.AXIS.Y:
-        cursorY += dimensions.height + gaps[i];
-        break;
-    }
-  }
-  this.workspace_.setResizesEnabled(true);
 };
 
 /**
@@ -1193,6 +1103,7 @@ Blockly.Flyout.prototype.layout_ = function(contents, gaps) {
   for (var i = 0, item; (item = contents[i]); i++) {
     if (item.type == 'block') {
       var block = item.block;
+      var root = block.getSvgRoot();
       var allBlocks = block.getDescendants(false);
       for (var j = 0, child; (child = allBlocks[j]); j++) {
         // Mark blocks as being inside a flyout.  This is used to detect and
@@ -1200,9 +1111,9 @@ Blockly.Flyout.prototype.layout_ = function(contents, gaps) {
         // block.
         child.isInFlyout = true;
       }
-      block.render();
       var blockHW = block.getHeightWidth();
       this.createRect_(block, 0, 0, blockHW, i);
+      this.addBlockListeners_(root, block, block.flyoutRect_);
       this.contents_.push(block);
     } else if (item.type == 'button') {
       this.initFlyoutButton_(item.button, 0, 0);
